pyproject.toml
[project]
name = "voter-api"
version = "0.1.0"
description = "Python/FastAPI REST API + CLI for managing Georgia Secretary of State voter data with geospatial capabilities"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "geoalchemy2>=0.15.0",
    "asyncpg>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "typer>=0.15.0",
    "loguru>=0.7.0",
    "pandas>=2.2.0",
    "alembic>=1.14.0",
    "pyjwt>=2.9.0",
    "httpx>=0.28.0",
    "geopandas>=1.0.0",
    "pyogrio>=0.10.0",
    "shapely>=2.0.0",
    "python-multipart>=0.0.18",
    "tqdm>=4.67.0",
    "email-validator>=2.3.0",
    "bcrypt>=5.0.0",
    "python-dateutil>=2.9.0.post0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pip-audit>=2.7.0",
    "aiosqlite>=0.22.1",
]

[project.scripts]
voter-api = "voter_api.cli.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voter_api"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra -q"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
target-version = "py313"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "S",   # flake8-bandit
    "A",   # flake8-builtins
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "T20", # flake8-print
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "PTH", # flake8-use-pathlib
]
ignore = [
    "S101",  # assert used (needed in tests)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S105", "S106"]
"src/voter_api/core/security.py" = ["S105", "S106"]
"src/voter_api/core/config.py" = ["S105", "S106"]
"src/voter_api/schemas/auth.py" = ["S105"]
"src/voter_api/api/**/*.py" = ["B008"]  # FastAPI Depends() in defaults is idiomatic
"src/voter_api/cli/**/*.py" = ["B008"]  # Typer Argument/Option in defaults is idiomatic

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.mypy.overrides]
module = [
    "geoalchemy2.*",
    "geopandas.*",
    "pyogrio.*",
    "shapely.*",
]
ignore_missing_imports = true
